Refactor day 1 exercises and built for linux

+++ src/WD1/ex13.c
@@ -1,11 +1,10 @@
 #include <stdio.h>
 
-int main()
-{
+int main() {
     int number;
     char letter;
     float pi;
-    long double value;
+    double value;
 
     number = 0;
     letter = 'a';
@@ -15,7 +14,7 @@ int main()
     printf("number :\t%d\n" , number);
     printf("letter :\t%c\n" , letter);
     printf("pi :\t\t%.2f\n" , pi);
-    printf("value :\t\t%.10Lf\n" , value);
+    printf("value :\t\t%.10lf\n" , value);
 
     return 0;
 }
